# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Meal Planner API",
    "description": "API for managing recipes, meal plans, and shopping lists",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/meal-plans": {
      "get": {
        "summary": "List all meal plans",
        "tags": [
          "Meal Plans"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "startDate": {
                        "type": "string"
                      },
                      "endDate": {
                        "type": "string"
                      },
                      "createdAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "days": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "dayDate": {
                              "type": "string"
                            },
                            "mealType": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "nullable": true,
                                  "enum": [
                                    null
                                  ]
                                }
                              ]
                            },
                            "recipe": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "number"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "description": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "nullable": true,
                                      "enum": [
                                        null
                                      ]
                                    }
                                  ]
                                },
                                "protein": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "nullable": true,
                                      "enum": [
                                        null
                                      ]
                                    }
                                  ]
                                },
                                "freezable": {
                                  "type": "boolean"
                                }
                              },
                              "required": [
                                "id",
                                "name",
                                "protein",
                                "freezable"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "required": [
                            "id",
                            "dayDate",
                            "mealType",
                            "recipe"
                          ],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "startDate",
                      "endDate",
                      "createdAt",
                      "days"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a meal plan",
        "tags": [
          "Meal Plans"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "startDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "endDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "days": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "dayDate": {
                          "type": "string",
                          "minLength": 1
                        },
                        "recipeId": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "maximum": 9007199254740991,
                          "minimum": 0
                        },
                        "mealType": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "dayDate",
                        "recipeId"
                      ]
                    }
                  }
                },
                "required": [
                  "name",
                  "startDate",
                  "endDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "startDate": {
                      "type": "string"
                    },
                    "endDate": {
                      "type": "string"
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "days": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "dayDate": {
                            "type": "string"
                          },
                          "mealType": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "recipe": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "nullable": true,
                                    "enum": [
                                      null
                                    ]
                                  }
                                ]
                              },
                              "protein": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "nullable": true,
                                    "enum": [
                                      null
                                    ]
                                  }
                                ]
                              },
                              "freezable": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "protein",
                              "freezable"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "required": [
                          "id",
                          "dayDate",
                          "mealType",
                          "recipe"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "startDate",
                    "endDate",
                    "createdAt",
                    "days"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/meal-plans/{id}": {
      "get": {
        "summary": "Get a meal plan by ID",
        "tags": [
          "Meal Plans"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "startDate": {
                      "type": "string"
                    },
                    "endDate": {
                      "type": "string"
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "days": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "dayDate": {
                            "type": "string"
                          },
                          "mealType": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "recipe": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "nullable": true,
                                    "enum": [
                                      null
                                    ]
                                  }
                                ]
                              },
                              "protein": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "nullable": true,
                                    "enum": [
                                      null
                                    ]
                                  }
                                ]
                              },
                              "freezable": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "protein",
                              "freezable"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "required": [
                          "id",
                          "dayDate",
                          "mealType",
                          "recipe"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "startDate",
                    "endDate",
                    "createdAt",
                    "days"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a meal plan",
        "tags": [
          "Meal Plans"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "startDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "endDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "days": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "dayDate": {
                          "type": "string",
                          "minLength": 1
                        },
                        "recipeId": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "maximum": 9007199254740991,
                          "minimum": 0
                        },
                        "mealType": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "dayDate",
                        "recipeId"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "startDate": {
                      "type": "string"
                    },
                    "endDate": {
                      "type": "string"
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "days": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "dayDate": {
                            "type": "string"
                          },
                          "mealType": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "recipe": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "nullable": true,
                                    "enum": [
                                      null
                                    ]
                                  }
                                ]
                              },
                              "protein": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "nullable": true,
                                    "enum": [
                                      null
                                    ]
                                  }
                                ]
                              },
                              "freezable": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "protein",
                              "freezable"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "required": [
                          "id",
                          "dayDate",
                          "mealType",
                          "recipe"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "startDate",
                    "endDate",
                    "createdAt",
                    "days"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a meal plan",
        "tags": [
          "Meal Plans"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}